Deploy 24/5

+++ app/Http/Controllers/ProfesorEscribeForo.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\usuarios;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Storage;
 use App\Models\datosForo;
 use App\Models\Foro;
@@ -22,7 +23,7 @@ class ProfesorEscribeForo extends Controller
             ->where('idGrupo', $idGrupo)->first());
     }
 
-    public function traerGrupos(Request $request,$id)
+    public function traerGrupos(Request $request, $id)
     {
         $usuario = usuarios::findOrFail($id);
         if ($usuario->ou == 'Profesor') {
@@ -53,22 +54,38 @@ class ProfesorEscribeForo extends Controller
         }
     }
 
-    public function show(Request $request)
+    public function getAllPublicaciones($idGrupo, $idUsuario, $limit)
     {
-        if ($request->idMateria) {
-            if ($request->ou == 'Profesor') {
-                return self::traerPublicacionesProfesorMateria($request);
-            } else if ($request->ou == 'Alumno') {
-                return self::traerPublicacionesAlumnoMateria($request);
-            }
-        } else {
-            if ($request->ou == 'Profesor') {
-                return self::traerPublicacionesProfesor($request);
-            } else if ($request->ou == 'Alumno') {
-                return self::traerPublicacionesAlumno($request);
-            }
+        $usuario = usuarios::findOrFail($idUsuario);
+        $data = [
+            'idUsuario' => $idUsuario,
+            'idGrupo' => $idGrupo,
+            'limit' => $limit
+        ];
+
+        if ($usuario->ou == 'Profesor') {
+            return self::traerPublicacionesProfesor($data);
+        } else if ($usuario->ou == 'Alumno') {
+            return self::traerPublicacionesAlumno($data);
         }
-        return response()->json(['error' => 'No se encontraron publicaciones'], 404);
+    }
+
+    public function getAllPublicacionesMateria($idGrupo, $idUsuario, $idMateria, $limit)
+    {
+        $usuario = usuarios::findOrFail($idUsuario);
+        $data = [
+            'idUsuario' => $idUsuario,
+            'idGrupo' => $idGrupo,
+            'idMateria' => $idMateria,
+            'limit' => $limit,
+        ];
+        if ($usuario->ou == 'Profesor') {
+            return self::traerPublicacionesProfesorMateria($data);
+        } else if ($usuario->ou == 'Alumno') {
+            return self::traerPublicacionesAlumnoMateria($data);
+        }
+
+        return response()->json([]);
     }
 
 
@@ -89,8 +106,12 @@ class ProfesorEscribeForo extends Controller
             $arrayImagenes = array();
             $postAuthor = $p->postAuthor;
             $imgPerfil = $this->getImagenPefil($postAuthor);
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            } else {
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
 
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
 
 
             foreach ($peticionSQLFiltrada as $p2) {
@@ -99,7 +120,12 @@ class ProfesorEscribeForo extends Controller
                 if ($resultado) {
                     array_push($arrayArchivos, $p2->archivo);
                 } else {
-                    array_push($arrayImagenes, base64_encode(Storage::disk('ftp')->get($p2->archivo)));
+                    if (!App::environment(['testing'])) {
+                        array_push($arrayImagenes, base64_encode(Storage::disk('ftp')->get($p2->archivo)));
+                    } else {
+                        array_push($arrayImagenes, $p2->archivo);
+                    }
+
                 }
             }
 
@@ -195,8 +221,12 @@ class ProfesorEscribeForo extends Controller
             $postAuthor = $p->postAuthor;
 
             $imgPerfil = $this->getImagenPefil($postAuthor);
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            } else {
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
 
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
 
             foreach ($peticionSQLFiltrada as $p2) {
 
@@ -247,8 +277,12 @@ class ProfesorEscribeForo extends Controller
             $postAuthor = $p->postAuthor;
 
             $imgPerfil = $this->getImagenPefil($postAuthor);
-
-            $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            if (!App::environment(['testing'])) {
+                $img = base64_encode(Storage::disk('ftp')->get($imgPerfil[0]->imagen_perfil));
+            }else{
+                $img = $imgPerfil[0]->imagen_perfil;
+            }
+          
 
             foreach ($peticionSQLFiltrada as $p2) {
 
@@ -287,13 +321,15 @@ class ProfesorEscribeForo extends Controller
     {
         $request->validate(
             [
-                'idForo' => 'required',
+                'idGrupo' => 'required',
+                'idMateria' => 'required',
                 'idUsuario' => 'required',
                 'mensaje' => 'string',
                 'archivos' => 'array | nullable',
                 'nombresArchivo' => 'array | nullable'
             ]
         );
+
 
         $datoForo = $this->agregarDatosForo($request);
         if ($request->archivos) {
@@ -308,8 +344,9 @@ class ProfesorEscribeForo extends Controller
 
     public function agregarDatosForo(Request $request)
     {
+        $foro = ProfesorForoGrupo::where('idGrupo', $request->idGrupo)->where('idMateria', $request->idMateria)->first();
         $datosForo = new datosForo;
-        $datosForo->idForo = $request->idForo;
+        $datosForo->idForo = $foro->idForo;
         $datosForo->idUsuario = $request->idUsuario;
         $datosForo->mensaje = $request->mensaje;
         $datosForo->save();
@@ -320,11 +357,13 @@ class ProfesorEscribeForo extends Controller
 
     public function destroy(Request $request, $id)
     {
-
         $postForo = datosForo::findOrFail($id);
         $arhivosForo = archivosForo::where('idDato', $id)->get();
         foreach ($arhivosForo as $p) {
-            Storage::disk('ftp')->delete($p->nombreArchivo);
+            if (!App::environment(['testing'])) {
+                Storage::disk('ftp')->delete($p->nombreArchivo);
+            }
+          
             $arhivosId = archivosForo::where('id', $p->id)->first();
             $arhivosId->delete();
         }
@@ -338,7 +377,7 @@ class ProfesorEscribeForo extends Controller
     }
 
 
-    public function getPublicacionesForoForProfesor($request): \Illuminate\Support\Collection
+    public function getPublicacionesForoForProfesor($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre AS materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
@@ -347,18 +386,18 @@ class ProfesorEscribeForo extends Controller
             ->join('grupos_tienen_profesor', 'grupos_tienen_profesor.idGrupo', '=', 'profesor_estan_grupo_foro.idGrupo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
-            ->where('profesor_estan_grupo_foro.idProfesor', $request->idUsuario)
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
+            ->where('profesor_estan_grupo_foro.idProfesor', $request['idUsuario'])
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
 
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->distinct()
             ->get();
         return $peticionSQL;
     }
 
 
-    public function getArchivosForo($p): \Illuminate\Support\Collection
+    public function getArchivosForo($p)
     {
         $peticionSQLFiltrada = DB::table('archivos_foro')
             ->select('nombreArchivo AS archivo')
@@ -369,7 +408,7 @@ class ProfesorEscribeForo extends Controller
     }
 
 
-    public function getImagenPefil($postAuthor): \Illuminate\Support\Collection
+    public function getImagenPefil($postAuthor)
     {
         $imgPerfil = DB::table('usuarios')
             ->select('imagen_perfil')
@@ -379,24 +418,24 @@ class ProfesorEscribeForo extends Controller
     }
 
 
-    public function getPublicacionesForoMateriaForProfesor($request): \Illuminate\Support\Collection
+    public function getPublicacionesForoMateriaForProfesor($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
-            ->where('profesor_estan_grupo_foro.idProfesor', $request->idUsuario)
-            ->where('profesor_estan_grupo_foro.idMateria', $request->idMateria)
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
+            ->where('profesor_estan_grupo_foro.idProfesor', $request['idUsuario'])
+            ->where('profesor_estan_grupo_foro.idMateria', $request['idMateria'])
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->get();
         return $peticionSQL;
     }
 
 
-    public function getIdGrupoAlumno($request): \Illuminate\Support\Collection
+    public function getIdGrupoAlumno($request)
     {
         $idGrupo = DB::table('alumnos_pertenecen_grupos')
             ->select('alumnos_pertenecen_grupos.idGrupo AS idGrupo')
@@ -406,33 +445,33 @@ class ProfesorEscribeForo extends Controller
     }
 
 
-    public function getPublicacionesForoForGrupo($request): \Illuminate\Support\Collection
+    public function getPublicacionesForoForGrupo($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->distinct()
             ->get();
         return $peticionSQL;
     }
 
 
-    public function getPublicacionesForoMateriaForGrupo($request): \Illuminate\Support\Collection
+    public function getPublicacionesForoMateriaForGrupo($request)
     {
         $peticionSQL = DB::table('profesor_estan_grupo_foro')
             ->select('datosForo.id AS id', 'datosForo.idForo AS idForo', 'profesor_estan_grupo_foro.idGrupo', 'materias.nombre as materia', 'datosForo.idUsuario AS idUsuario', 'usuarios.nombre AS nombreAutor', 'datosForo.mensaje AS mensaje', 'datosForo.created_at AS fecha', 'datosForo.idUsuario as postAuthor')
             ->join('materias', 'materias.id', '=', 'profesor_estan_grupo_foro.idMateria')
             ->join('datosForo', 'datosForo.idForo', '=', 'profesor_estan_grupo_foro.idForo')
             ->join('usuarios', 'usuarios.id', '=', 'datosForo.idUsuario')
-            ->where('profesor_estan_grupo_foro.idGrupo', $request->idGrupo)
-            ->where('profesor_estan_grupo_foro.idMateria', $request->idMateria)
+            ->where('profesor_estan_grupo_foro.idGrupo', $request['idGrupo'])
+            ->where('profesor_estan_grupo_foro.idMateria', $request['idMateria'])
             ->orderBy('id', 'desc')
-            ->take($request->limit)
+            ->take($request['limit'])
             ->get();
         return $peticionSQL;
     }
@@ -443,11 +482,15 @@ class ProfesorEscribeForo extends Controller
 
     public function subirArchivoForo(Request $request, int $i, $idDatos)
     {
+        $foro = ProfesorForoGrupo::where('idGrupo', $request->idGrupo)->where('idMateria', $request->idMateria)->first();
         $nombreArchivo = random_int(0, 1000000) . "_" . $request->nombresArchivo[$i];
-        Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        if (!App::environment(['testing'])) {
+            Storage::disk('ftp')->put($nombreArchivo, fopen($request->archivos[$i], 'r+'));
+        }
+       
         $archivosForo = new archivosForo;
-        $archivosForo->idDato = $idDatos->id;
-        $archivosForo->idForo = $request->idForo;
+        $archivosForo->idDato = $idDatos;
+        $archivosForo->idForo = $foro->idForo;
         $archivosForo->nombreArchivo = $nombreArchivo;
         $archivosForo->save();
     }
